Prevent game from moving the freecam
Previous version was not able to consistently prevent the game engine from endlessly moving the freecam, so it disabled mouse support if the player is using mouse movement.  Now it invalidates large sweeping movements (>1/4 the width/height of the screen), so at least the screen doesn't go crazy when the game engine refuses to release control of the mouse.  The pan/tilt keys have also been changed since the game already uses the arrow keys.